@@ -1,17 +1,33 @@
 {
-  "require": {
-    "php": ">=8.0",
-    "ext-exif": "*",
-    "psr/log": "^3.0",
-    "symfony/dependency-injection": "^6.0"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "^9.6"
-  },
+  "name": "yoannblot/photos-mover",
+  "description": "CLI application that automatically organizes photos and videos by their creation date",
+  "type": "project",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Yoann Blot"
+    }
+  ],
   "autoload": {
     "psr-4": {
       "App\\": "src/",
       "Tests\\": "tests/"
+    }
+  },
+  "require": {
+    "php": ">=8.4",
+    "ext-exif": "*",
+    "symfony/dependency-injection": "^8.0.2",
+    "psr/log": "^3.0.2"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^12.5.4",
+    "rector/rector": "^2.3",
+    "carthage-software/mago": "^1.0.3"
+  },
+  "config": {
+    "allow-plugins": {
+      "carthage-software/mago": true
     }
   }
 }
