@@ -8,34 +8,34 @@ use App\Domain\Finder;
 use App\Domain\Mover;
 use App\Domain\PathGenerator;
 use App\Domain\Type\Directory;
+use InvalidArgumentException;
 use Psr\Log\LoggerInterface;
 
-final class MoveMediaFiles
+final readonly class MoveMediaFiles
 {
-    private Finder $finder;
-    private PathGenerator $pathGenerator;
-    private Mover $mover;
-    private LoggerInterface $logger;
-
-    public function __construct(Finder $finder, PathGenerator $pathGenerator, Mover $mover, LoggerInterface $logger)
-    {
-        $this->finder = $finder;
-        $this->pathGenerator = $pathGenerator;
-        $this->mover = $mover;
-        $this->logger = $logger;
-    }
+    public function __construct(
+        private Finder $finder,
+        private PathGenerator $pathGenerator,
+        private Mover $mover,
+        private LoggerInterface $logger,
+    ) {}
 
     public function move(Directory $source, Directory $destination): void
     {
-        $this->logger->info("Start moving files from '{$source->getPath()}' to '{$destination->getPath()}'");
-        foreach ($this->finder->find($source) as $file) {
-            $this->logger->debug("Check for '{$file->getPath()}'");
+        $this->logger->info(sprintf(
+            "Start moving files from '%s' to '%s'",
+            $source->getPath(),
+            $destination->getPath(),
+        ));
+        foreach ($this->finder->find($source) as $file)
+        {
+            $this->logger->debug(sprintf("Check for '%s'", $file->getPath()));
             try {
                 $newFilePath = $this->pathGenerator->generate($destination, $file);
-                $this->logger->info("Will move '{$file->getPath()}' to $newFilePath'");
+                $this->logger->info(sprintf("Will move '%s' to %s'", $file->getPath(), $newFilePath));
                 $this->mover->move($file, $newFilePath);
-            } catch (\InvalidArgumentException $e) {
-                $this->logger->warning("Unable to move file: '{$file->getPath()}'");
+            } catch (InvalidArgumentException) {
+                $this->logger->warning(sprintf("Unable to move file: '%s'", $file->getPath()));
             }
         }
     }
