@@ -8,34 +8,35 @@ use App\Domain\Type\File;
 use Psr\Log\LoggerInterface;
 use RuntimeException;
 
-final class Mover
+final readonly class Mover
 {
-    private LoggerInterface $logger;
+    public function __construct(
+        private LoggerInterface $logger,
+    ) {}
 
-    public function __construct(LoggerInterface $logger)
-    {
-        $this->logger = $logger;
-    }
-
-    public function move(File $source, string $destinationPath): void
+    public function move(File $file, string $destinationPath): void
     {
         $this->createDirectoryIfNecessary($destinationPath);
 
-        if (rename($source->getPath(), $destinationPath)) {
-            $this->logger->debug("Move '{$source->getPath()}' to '$destinationPath'");
-        } else {
-            $this->logger->error("Unable to move '{$source->getPath()}' to '$destinationPath'");
+        if (rename($file->getPath(), $destinationPath))
+        {
+            $this->logger->debug(sprintf("Move '%s' to '%s'", $file->getPath(), $destinationPath));
+        } else
+        {
+            $this->logger->error(sprintf("Unable to move '%s' to '%s'", $file->getPath(), $destinationPath));
         }
     }
 
     private function createDirectoryIfNecessary(string $destination): void
     {
         $destinationDirectory = dirname($destination);
-        if (is_dir($destinationDirectory)) {
+        if (is_dir($destinationDirectory))
+        {
             return;
         }
 
-        if (!mkdir($destinationDirectory, 0705, true) && !is_dir($destinationDirectory)) {
+        if (!mkdir($destinationDirectory, 0o705, true) && !is_dir($destinationDirectory))
+        {
             throw new RuntimeException(sprintf('Directory "%s" was not created', $destinationDirectory));
         }
     }
