@@ -8,19 +8,16 @@ use App\Domain\Metadata\FileMetadataReader;
 use App\Domain\Type\Directory;
 use App\Domain\Type\File;
 
-final class PathGenerator
+final readonly class PathGenerator
 {
-    private FileMetadataReader $fileMetadataReader;
-
-    public function __construct(FileMetadataReader $fileMetadataReader)
-    {
-        $this->fileMetadataReader = $fileMetadataReader;
-    }
+    public function __construct(
+        private FileMetadataReader $fileMetadataReader,
+    ) {}
 
     public function generate(Directory $outputDirectory, File $file): string
     {
-        $metadata = $this->fileMetadataReader->extractMetadata($file);
-        $newDirectory = $outputDirectory->getPath() . $metadata->getDate()->format('Y/m/d') . DIRECTORY_SEPARATOR;
+        $fileMetadata = $this->fileMetadataReader->extractMetadata($file);
+        $newDirectory = $outputDirectory->getPath() . $fileMetadata->getDate()->format('Y/m/d') . DIRECTORY_SEPARATOR;
 
         return $newDirectory . strtolower($file->getDirectory());
     }
