@@ -8,18 +8,29 @@ use App\Domain\Metadata\FileReader;
 use App\Domain\Type\File;
 use App\Domain\Type\FileMetadata;
 use DateTimeImmutable;
+use InvalidArgumentException;
 
 final class DefaultFileMetadataReader implements FileReader
 {
-    public function supports(File $file): bool
-    {
-        return $file->isImage() || $file->isVideo();
-    }
-
     public function extractMetadata(File $file): FileMetadata
     {
         $timestamp = filemtime($file->getPath());
+        $date      = DateTimeImmutable::createFromFormat('U', (string) $timestamp);
+        if ($date === false)
+        {
+            throw new InvalidArgumentException(sprintf("Unable to extract metadata from file: '%s'", $file->getPath()));
+        }
 
-        return new FileMetadata(DateTimeImmutable::createFromFormat('U', (string) $timestamp));
+        return new FileMetadata($date);
+    }
+
+    public function supports(File $file): bool
+    {
+        if ($file->isImage())
+        {
+            return true;
+        }
+
+        return $file->isVideo();
     }
 }
