@@ -4,12 +4,54 @@ declare(strict_types=1);
 
 namespace App\Infrastructure;
 
-use Psr\Log\AbstractLogger;
+use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
+use Stringable;
 
-final class StdoutLogger extends AbstractLogger
+final class StdoutLogger implements LoggerInterface
 {
-    public function log($level, \Stringable|string $message, array $context = []): void
+    public function alert(string|Stringable $message, array $context = []): void
     {
-        printf("[%s] %s" . PHP_EOL, $level, $message);
+        $this->log(LogLevel::ALERT, $message, $context);
+    }
+
+    public function critical(string|Stringable $message, array $context = []): void
+    {
+        $this->log(LogLevel::CRITICAL, $message, $context);
+    }
+
+    public function debug(string|Stringable $message, array $context = []): void
+    {
+        $this->log(LogLevel::DEBUG, $message, $context);
+    }
+
+    public function emergency(string|Stringable $message, array $context = []): void
+    {
+        $this->log(LogLevel::EMERGENCY, $message, $context);
+    }
+
+    public function error(string|Stringable $message, array $context = []): void
+    {
+        $this->log(LogLevel::ERROR, $message, $context);
+    }
+
+    public function info(string|Stringable $message, array $context = []): void
+    {
+        $this->log(LogLevel::INFO, $message, $context);
+    }
+
+    public function log(mixed $level, Stringable|string $message, array $context = []): void
+    {
+        printf('[%s] %s' . PHP_EOL, (string) $level, $message);
+    }
+
+    public function notice(string|Stringable $message, array $context = []): void
+    {
+        $this->log(LogLevel::NOTICE, $message, $context);
+    }
+
+    public function warning(string|Stringable $message, array $context = []): void
+    {
+        $this->log(LogLevel::WARNING, $message, $context);
     }
 }
