@@ -9,40 +9,40 @@ use App\Domain\Type\File;
 
 final class Fixtures
 {
-    private const FIXTURES_DIRECTORY = __DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'Fixtures';
+    private const string FIXTURES_DIRECTORY = __DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'Fixtures';
 
-    public static function duplicateJpgImageIn(DirectoryType $destination): void
+    public static function createJpgImageFile(DirectoryType $directoryType): File
     {
-        copy(self::getJpgImageFile()->getPath(), $destination->getPath() . DIRECTORY_SEPARATOR . 'image.jpg');
-    }
-
-    public static function duplicatePngImageIn(DirectoryType $destination): void
-    {
-        copy(self::getPngImageFile()->getPath(), $destination->getPath() . DIRECTORY_SEPARATOR . 'image.png');
-    }
-
-    public static function duplicateVideoIn(DirectoryType $destination): void
-    {
-        copy(
-            self::getVideoFileWithDateTimeInName()->getPath(),
-            $destination->getPath() . DIRECTORY_SEPARATOR . self::getVideoFileWithDateTimeInName()->getFileName()
-        );
-    }
-
-    public static function createJpgImageFile(DirectoryType $destination): File
-    {
-        $imagePath = $destination->getPath() . DIRECTORY_SEPARATOR . 'image.jpg';
+        $imagePath = $directoryType->getPath() . DIRECTORY_SEPARATOR . 'image.jpg';
         copy(self::getJpgImageFile()->getPath(), $imagePath);
 
         return new File($imagePath);
     }
 
-    public static function createPngImageFile(DirectoryType $destination): File
+    public static function createPngImageFile(DirectoryType $directoryType): File
     {
-        $imagePath = $destination->getPath() . DIRECTORY_SEPARATOR . 'image.png';
+        $imagePath = $directoryType->getPath() . DIRECTORY_SEPARATOR . 'image.png';
         copy(self::getPngImageFile()->getPath(), $imagePath);
 
         return new File($imagePath);
+    }
+
+    public static function duplicateJpgImageIn(DirectoryType $directoryType): void
+    {
+        copy(self::getJpgImageFile()->getPath(), $directoryType->getPath() . DIRECTORY_SEPARATOR . 'image.jpg');
+    }
+
+    public static function duplicatePngImageIn(DirectoryType $directoryType): void
+    {
+        copy(self::getPngImageFile()->getPath(), $directoryType->getPath() . DIRECTORY_SEPARATOR . 'image.png');
+    }
+
+    public static function duplicateVideoIn(DirectoryType $directoryType): void
+    {
+        copy(
+            self::getVideoFileWithDateTimeInName()->getPath(),
+            $directoryType->getPath() . DIRECTORY_SEPARATOR . self::getVideoFileWithDateTimeInName()->getFileName(),
+        );
     }
 
     public static function getJpgImageFile(): File
@@ -55,6 +55,11 @@ final class Fixtures
         return new File(self::FIXTURES_DIRECTORY . DIRECTORY_SEPARATOR . 'image.png');
     }
 
+    public static function getTextFile(): File
+    {
+        return new File(self::FIXTURES_DIRECTORY . DIRECTORY_SEPARATOR . 'text.txt');
+    }
+
     public static function getVideoFile(): File
     {
         return new File(self::FIXTURES_DIRECTORY . DIRECTORY_SEPARATOR . 'video.mp4');
@@ -63,10 +68,5 @@ final class Fixtures
     public static function getVideoFileWithDateTimeInName(): File
     {
         return new File(self::FIXTURES_DIRECTORY . DIRECTORY_SEPARATOR . 'VID20230731221612.mp4');
-    }
-
-    public static function getTextFile(): File
-    {
-        return new File(self::FIXTURES_DIRECTORY . DIRECTORY_SEPARATOR . 'text.txt');
     }
 }
