@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Tests\Integration\Application;
 
 use App\Application\MoveMediaFiles;
+use DateTimeImmutable;
+use Override;
 use Tests\Helper\Directory as DirectoryHelper;
 use Tests\Helper\Fixtures;
 use Tests\Integration\IntegrationTestCase;
@@ -13,17 +15,10 @@ final class MoveMediaFilesTest extends IntegrationTestCase
 {
     private MoveMediaFiles $sut;
 
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $this->sut = $this->app->get(MoveMediaFiles::class);
-    }
-
-    /** @test */
-    public function it_moves_a_jpg_image_based_on_its_metadata(): void
+    public function test_it_moves_a_jpg_image_based_on_its_metadata(): void
     {
         // Arrange
-        $sourceDirectory = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
+        $sourceDirectory      = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
         $destinationDirectory = DirectoryHelper::create('Output-' . __FUNCTION__);
         Fixtures::duplicateJpgImageIn($sourceDirectory);
 
@@ -31,11 +26,15 @@ final class MoveMediaFilesTest extends IntegrationTestCase
         $this->sut->move($sourceDirectory, $destinationDirectory);
 
         // Assert
-        $today = new \DateTimeImmutable();
-        $expectedImagePath = $destinationDirectory->getPath() . $today->format('Y')
-            . DIRECTORY_SEPARATOR . $today->format('m')
-            . DIRECTORY_SEPARATOR . $today->format('d')
-            . DIRECTORY_SEPARATOR . 'image.jpg';
+        $today             = new DateTimeImmutable();
+        $expectedImagePath = $destinationDirectory->getPath()
+        . $today->format('Y')
+        . DIRECTORY_SEPARATOR
+        . $today->format('m')
+        . DIRECTORY_SEPARATOR
+        . $today->format('d')
+        . DIRECTORY_SEPARATOR
+        . 'image.jpg';
         $this->assertFileExists($expectedImagePath);
 
         unlink($expectedImagePath);
@@ -43,11 +42,10 @@ final class MoveMediaFilesTest extends IntegrationTestCase
         DirectoryHelper::remove($sourceDirectory);
     }
 
-    /** @test */
-    public function it_moves_a_png_image_based_on_its_metadata(): void
+    public function test_it_moves_a_png_image_based_on_its_metadata(): void
     {
         // Arrange
-        $sourceDirectory = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
+        $sourceDirectory      = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
         $destinationDirectory = DirectoryHelper::create('Output-' . __FUNCTION__);
         Fixtures::duplicatePngImageIn($sourceDirectory);
 
@@ -55,11 +53,15 @@ final class MoveMediaFilesTest extends IntegrationTestCase
         $this->sut->move($sourceDirectory, $destinationDirectory);
 
         // Assert
-        $today = new \DateTimeImmutable();
-        $expectedImagePath = $destinationDirectory->getPath() . $today->format('Y')
-            . DIRECTORY_SEPARATOR . $today->format('m')
-            . DIRECTORY_SEPARATOR . $today->format('d')
-            . DIRECTORY_SEPARATOR . 'image.png';
+        $today             = new DateTimeImmutable();
+        $expectedImagePath = $destinationDirectory->getPath()
+        . $today->format('Y')
+        . DIRECTORY_SEPARATOR
+        . $today->format('m')
+        . DIRECTORY_SEPARATOR
+        . $today->format('d')
+        . DIRECTORY_SEPARATOR
+        . 'image.png';
         $this->assertFileExists($expectedImagePath);
 
         unlink($expectedImagePath);
@@ -67,11 +69,10 @@ final class MoveMediaFilesTest extends IntegrationTestCase
         DirectoryHelper::remove($sourceDirectory);
     }
 
-    /** @test */
-    public function it_moves_a_video_based_on_its_metadata(): void
+    public function test_it_moves_a_video_based_on_its_metadata(): void
     {
         // Arrange
-        $sourceDirectory = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
+        $sourceDirectory      = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
         $destinationDirectory = DirectoryHelper::create('Output-' . __FUNCTION__);
         Fixtures::duplicateVideoIn($sourceDirectory);
 
@@ -79,14 +80,26 @@ final class MoveMediaFilesTest extends IntegrationTestCase
         $this->sut->move($sourceDirectory, $destinationDirectory);
 
         // Assert
-        $expectedVideoPath = $destinationDirectory->getPath() . '2023'
-            . DIRECTORY_SEPARATOR . '07'
-            . DIRECTORY_SEPARATOR . '31'
-            . DIRECTORY_SEPARATOR . 'vid20230731221612.mp4';
+        $expectedVideoPath =
+            $destinationDirectory->getPath()
+            . '2023'
+            . DIRECTORY_SEPARATOR
+            . '07'
+            . DIRECTORY_SEPARATOR
+            . '31'
+            . DIRECTORY_SEPARATOR
+            . 'vid20230731221612.mp4';
         $this->assertFileExists($expectedVideoPath);
 
         unlink($expectedVideoPath);
         DirectoryHelper::remove($destinationDirectory);
         DirectoryHelper::remove($sourceDirectory);
+    }
+
+    #[Override]
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->sut = $this->app->get(MoveMediaFiles::class);
     }
 }
