@@ -10,6 +10,7 @@ use App\Domain\Metadata\FileMetadataReader;
 use App\Domain\Mover;
 use App\Domain\PathGenerator;
 use App\Infrastructure\Metadata\ExifMetadataReader;
+use DateTimeImmutable;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
 use Tests\Helper\Directory as DirectoryHelper;
@@ -19,26 +20,10 @@ final class MoveMediaFilesTest extends TestCase
 {
     private MoveMediaFiles $sut;
 
-    protected function setUp(): void
-    {
-        $logger = new NullLogger();
-        $this->sut = new MoveMediaFiles(
-            new Finder(),
-            new PathGenerator(
-                new FileMetadataReader([
-                    new ExifMetadataReader(),
-                ])
-            ),
-            new Mover($logger),
-            $logger,
-        );
-    }
-
-    /** @test */
-    public function it_moves_an_image_based_on_its_metadata(): void
+    public function test_it_moves_an_image_based_on_its_metadata(): void
     {
         // Arrange
-        $sourceDirectory = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
+        $sourceDirectory      = DirectoryHelper::create('Fixtures-' . __FUNCTION__);
         $destinationDirectory = DirectoryHelper::create('Output-' . __FUNCTION__);
         Fixtures::duplicateJpgImageIn($sourceDirectory);
 
@@ -46,15 +31,32 @@ final class MoveMediaFilesTest extends TestCase
         $this->sut->move($sourceDirectory, $destinationDirectory);
 
         // Assert
-        $today = new \DateTimeImmutable();
-        $expectedImagePath = $destinationDirectory->getPath() . $today->format('Y')
-            . DIRECTORY_SEPARATOR . $today->format('m')
-            . DIRECTORY_SEPARATOR . $today->format('d')
-            . DIRECTORY_SEPARATOR . 'image.jpg';
+        $today             = new DateTimeImmutable();
+        $expectedImagePath = $destinationDirectory->getPath()
+        . $today->format('Y')
+        . DIRECTORY_SEPARATOR
+        . $today->format('m')
+        . DIRECTORY_SEPARATOR
+        . $today->format('d')
+        . DIRECTORY_SEPARATOR
+        . 'image.jpg';
         $this->assertFileExists($expectedImagePath);
 
         unlink($expectedImagePath);
         DirectoryHelper::remove($destinationDirectory);
         DirectoryHelper::remove($sourceDirectory);
+    }
+
+    protected function setUp(): void
+    {
+        $nullLogger = new NullLogger();
+        $this->sut = new MoveMediaFiles(
+            new Finder(),
+            new PathGenerator(new FileMetadataReader([
+                new ExifMetadataReader(),
+            ])),
+            new Mover($nullLogger),
+            $nullLogger,
+        );
     }
 }
