@@ -10,50 +10,15 @@ use FilesystemIterator;
 use PHPUnit\Framework\TestCase;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
+use SplFileInfo;
 
 final class FinderTest extends TestCase
 {
     private Finder $sut;
+
     private Directory $directory;
 
-    protected function setUp(): void
-    {
-        $this->sut = new Finder();
-        $directoryPath = __DIR__ . DIRECTORY_SEPARATOR . 'tmp';
-        mkdir($directoryPath);
-        $this->directory = new Directory($directoryPath);
-    }
-
-    protected function tearDown(): void
-    {
-        // remove all files in directory
-        $di = new RecursiveDirectoryIterator($this->directory->getPath(), FilesystemIterator::SKIP_DOTS);
-        $ri = new RecursiveIteratorIterator($di, RecursiveIteratorIterator::CHILD_FIRST);
-        foreach ($ri as $file) {
-            /** @var \SplFileInfo $file */
-            $file->isDir() ? rmdir($file->getRealPath()) : unlink($file->getRealPath());
-        }
-        rmdir($this->directory->getPath());
-    }
-
-    /** @test */
-    public function it_finds_an_image(): void
-    {
-        // Arrange
-        $this->createFile('test.jpg');
-
-        // Act
-        $files = $this->sut->find($this->directory);
-
-        // Assert
-        $this->assertCount(1, $files);
-        $file = $files[0];
-        $this->assertSame('test.jpg', $file->getFileName());
-        $this->assertTrue($file->isImage());
-    }
-
-    /** @test */
-    public function it_finds_a_video(): void
+    public function test_it_finds_a_video(): void
     {
         // Arrange
         $this->createFile('test.mp4');
@@ -68,8 +33,22 @@ final class FinderTest extends TestCase
         $this->assertTrue($file->isVideo());
     }
 
-    /** @test */
-    public function it_finds_files_only_in_root_directory(): void
+    public function test_it_finds_an_image(): void
+    {
+        // Arrange
+        $this->createFile('test.jpg');
+
+        // Act
+        $files = $this->sut->find($this->directory);
+
+        // Assert
+        $this->assertCount(1, $files);
+        $file = $files[0];
+        $this->assertSame('test.jpg', $file->getFileName());
+        $this->assertTrue($file->isImage());
+    }
+
+    public function test_it_finds_files_only_in_root_directory(): void
     {
         // Arrange
         $this->createFile('test.jpg');
@@ -82,6 +61,33 @@ final class FinderTest extends TestCase
 
         // Assert
         $this->assertCount(2, $files);
+    }
+
+    protected function setUp(): void
+    {
+        $this->sut = new Finder();
+        $directoryPath = __DIR__ . DIRECTORY_SEPARATOR . 'tmp';
+        mkdir($directoryPath);
+        $this->directory = new Directory($directoryPath);
+    }
+
+    protected function tearDown(): void
+    {
+        $di = new RecursiveDirectoryIterator($this->directory->getPath(), FilesystemIterator::SKIP_DOTS);
+        $ri = new RecursiveIteratorIterator($di, RecursiveIteratorIterator::CHILD_FIRST);
+        foreach ($ri as $file)
+        {
+            /** @var SplFileInfo $file */
+            $filePath = $file->getRealPath();
+            if ($filePath === false)
+            {
+                continue;
+            }
+
+            $file->isDir() ? rmdir($filePath) : unlink($filePath);
+        }
+
+        rmdir($this->directory->getPath());
     }
 
     private function createFile(string $fileName): void
