@@ -13,16 +13,10 @@ final class MoverTest extends TestCase
 {
     private Mover $sut;
 
-    protected function setUp(): void
-    {
-        $this->sut = new Mover(new NullLogger());
-    }
-
-    /** @test */
-    public function it_creates_the_destination_directory_if_it_does_not_exist(): void
+    public function test_it_creates_the_destination_directory_if_it_does_not_exist(): void
     {
         // Arrange
-        $sourcePath = __DIR__ . DIRECTORY_SEPARATOR . 'source-file.txt';
+        $sourcePath      = __DIR__ . DIRECTORY_SEPARATOR . 'source-file.txt';
         $destinationPath = __DIR__ . DIRECTORY_SEPARATOR . 'destination' . DIRECTORY_SEPARATOR . 'file.txt';
         file_put_contents($sourcePath, 'test');
 
@@ -35,11 +29,10 @@ final class MoverTest extends TestCase
         rmdir(dirname($destinationPath));
     }
 
-    /** @test */
-    public function it_moves_a_file_into_a_directory(): void
+    public function test_it_moves_a_file_into_a_directory(): void
     {
         // Arrange
-        $sourcePath = __DIR__ . DIRECTORY_SEPARATOR . 'source-file.txt';
+        $sourcePath      = __DIR__ . DIRECTORY_SEPARATOR . 'source-file.txt';
         $destinationPath = __DIR__ . DIRECTORY_SEPARATOR . 'destination-file.txt';
         file_put_contents($sourcePath, 'test');
 
@@ -50,5 +43,10 @@ final class MoverTest extends TestCase
         $this->assertFileExists($destinationPath);
         $this->assertFileDoesNotExist($sourcePath);
         unlink($destinationPath);
+    }
+
+    protected function setUp(): void
+    {
+        $this->sut = new Mover(new NullLogger());
     }
 }
