@@ -5,24 +5,13 @@ declare(strict_types=1);
 namespace Unit\Domain\Type;
 
 use App\Domain\Type\File;
+use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Tests\Helper\Fixtures;
 
 final class FileTest extends TestCase
 {
-    /** @test */
-    public function it_throws_an_invalid_argument_exception_when_file_does_not_exist(): void
-    {
-        // Assert
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage("File 'fake-path' does not exist.");
-
-        // Arrange & Act
-        new File('fake-path');
-    }
-
-    /** @test */
-    public function it_retrieves_directory_name(): void
+    public function test_it_retrieves_directory_name(): void
     {
         // Arrange
         $file = new File(Fixtures::getTextFile()->getPath());
@@ -34,8 +23,19 @@ final class FileTest extends TestCase
         $this->assertSame('text.txt', $directoryName);
     }
 
-    /** @test */
-    public function it_returns_true_when_file_is_an_image(): void
+    public function test_it_returns_false_when_file_is_not_an_image(): void
+    {
+        // Arrange
+        $file = new File(Fixtures::getVideoFile()->getPath());
+
+        // Act
+        $isImage = $file->isImage();
+
+        // Assert
+        $this->assertFalse($isImage);
+    }
+
+    public function test_it_returns_true_when_file_is_an_image(): void
     {
         // Arrange
         $file = new File(Fixtures::getJpgImageFile()->getPath());
@@ -47,16 +47,13 @@ final class FileTest extends TestCase
         $this->assertTrue($isImage);
     }
 
-    /** @test */
-    public function it_returns_false_when_file_is_not_an_image(): void
+    public function test_it_throws_an_invalid_argument_exception_when_file_does_not_exist(): void
     {
-        // Arrange
-        $file = new File(Fixtures::getVideoFile()->getPath());
+        // Assert
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage("File 'fake-path' does not exist.");
 
         // Act
-        $isImage = $file->isImage();
-
-        // Assert
-        $this->assertFalse($isImage);
+        new File('fake-path');
     }
 }
